Move per crate macros test to another dir
